+++ test/unit/readmeExampleResult.hs
@@ -1,0 +1,7 @@
+(
+  fromList [Node 0,Node 1],
+  fromList [(Node 0,0),(Node 1,1)],
+  fromList [(Node 1,Node 0,3)],
+  fromList [Node 1],
+  fromList [Node 0]
+  )